Factor animation export/playback helpers and fix build includes (#557)
