+++ src/controller/controls/ControlAnimation.cpp
@@ -1,6 +1,7 @@
 #include "controller/controls/ControlAnimation.h"
 #include "controller/Controller.h"
 #include "controller/ControllerContext.h"
+#include "controller/controls/AnimationHelper.h"
 #include "models/graph/AGraphNode.h"
 #include "models/graph/GraphManager.h"
 #include "models/graph/CameraNode.h"
@@ -126,69 +127,6 @@ namespace control::animation
             return ordered;
         }
 
-        std::vector<SafePtr<ViewPointNode>> collectPerspectiveViewpointsSorted(Controller& controller)
-        {
-            std::vector<SafePtr<ViewPointNode>> viewpoints;
-            const std::unordered_set<SafePtr<AGraphNode>> allViewpoints = controller.getGraphManager().getNodesByTypes({ ElementType::ViewPoint }, ObjectStatusFilter::ALL);
-            viewpoints.reserve(allViewpoints.size());
-
-            for (const SafePtr<AGraphNode>& node : allViewpoints)
-            {
-                SafePtr<ViewPointNode> viewpoint = static_pointer_cast<ViewPointNode>(node);
-                ReadPtr<ViewPointNode> rViewpoint = viewpoint.cget();
-                if (!rViewpoint || rViewpoint->getProjectionMode() != ProjectionMode::Perspective)
-                    continue;
-                viewpoints.push_back(viewpoint);
-            }
-
-            std::sort(viewpoints.begin(), viewpoints.end(), [](const SafePtr<ViewPointNode>& a, const SafePtr<ViewPointNode>& b)
-                {
-                    ReadPtr<ViewPointNode> ra = a.cget();
-                    ReadPtr<ViewPointNode> rb = b.cget();
-                    if (!ra || !rb)
-                        return bool(ra);
-                    return ra->getUserIndex() < rb->getUserIndex();
-                });
-
-            return viewpoints;
-        }
-
-        bool areViewpointsInterpolationCompatible(const ViewPointNode& reference, const ViewPointNode& candidate)
-        {
-            return reference.m_mode == candidate.m_mode &&
-                reference.m_blendMode == candidate.m_blendMode &&
-                reference.m_reduceFlash == candidate.m_reduceFlash &&
-                reference.m_flashAdvanced == candidate.m_flashAdvanced &&
-                reference.m_negativeEffect == candidate.m_negativeEffect &&
-                reference.m_postRenderingNormals.show == candidate.m_postRenderingNormals.show &&
-                reference.m_postRenderingAmbientOcclusion.enabled == candidate.m_postRenderingAmbientOcclusion.enabled &&
-                reference.m_postRenderingNormals.blendColor == candidate.m_postRenderingNormals.blendColor &&
-                reference.m_edgeAwareBlur.enabled == candidate.m_edgeAwareBlur.enabled &&
-                reference.m_depthLining.enabled == candidate.m_depthLining.enabled &&
-                reference.m_depthLining.strongMode == candidate.m_depthLining.strongMode;
-        }
-
-        bool canInterpolateSelectedViewpoints(const std::vector<SafePtr<ViewPointNode>>& viewpoints)
-        {
-            if (viewpoints.size() < 2)
-                return false;
-
-            ReadPtr<ViewPointNode> rReference = viewpoints.front().cget();
-            if (!rReference)
-                return false;
-
-            for (size_t i = 1; i < viewpoints.size(); ++i)
-            {
-                ReadPtr<ViewPointNode> rCandidate = viewpoints[i].cget();
-                if (!rCandidate)
-                    return false;
-
-                if (!areViewpointsInterpolationCompatible(*&rReference, *&rCandidate))
-                    return false;
-            }
-
-            return true;
-        }
     }
 
     AddViewPoint::AddViewPoint(SafePtr<AGraphNode> toAdd)
@@ -288,7 +226,7 @@ namespace control::animation
         }
 
         std::unordered_map<xg::Guid, SafePtr<ViewPointNode>> perspectiveById;
-        for (const SafePtr<ViewPointNode>& viewpoint : collectPerspectiveViewpointsSorted(controller))
+        for (const SafePtr<ViewPointNode>& viewpoint : helper::collectPerspectiveViewpointsSorted(controller))
         {
             ReadPtr<ViewPointNode> rViewpoint = viewpoint.cget();
             if (!rViewpoint)
@@ -342,7 +280,7 @@ namespace control::animation
         bool enableInterpolation = false;
         if (m_interpolateRenderingBetweenViewpoints)
         {
-            enableInterpolation = canInterpolateSelectedViewpoints(viewpoints);
+            enableInterpolation = helper::canInterpolateSelectedViewpoints(viewpoints);
             if (!enableInterpolation)
                 controller.updateInfo(new GuiDataWarning(TEXT_CONTEXT_ANIMATION_INCONSISTENT_INTERPOLATION));
         }
